reorganize library
